apcs_agent_system/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt --python-platform linux
agno==1.8.4
agno-aws==0.0.1
agno-docker==0.0.1
aiofiles==24.1.0
alembic==1.15.1
annotated-types==0.7.0
anyio==4.9.0
attrs==25.4.0
beautifulsoup4==4.13.3
boto3==1.37.19
botocore==1.37.19
certifi==2025.1.31
cffi==2.0.0
charset-normalizer==3.4.1
click==8.1.8
cryptography==46.0.4
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
docstring-parser==0.16
duckduckgo-search==7.5.3
email-validator==2.2.0
eval-type-backport==0.3.1
fastapi==0.115.12
fastapi-cli==0.0.7
feedparser==6.0.11
filelock==3.18.0
frozendict==2.4.6
gitdb==4.0.12
gitpython==3.1.44
googleapis-common-protos==1.72.0
googlesearch-python==1.3.0
greenlet==3.3.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.3
idna==3.10
importlib-metadata==8.7.1
invoke==2.2.1
jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lxml==5.3.1
lxml-html-clean==0.4.2
mako==1.3.9
markdown-it-py==3.0.0
markupsafe==3.0.2
mcp==1.26.0
mdurl==0.1.2
mistralai==1.12.0
multitasking==0.0.11
newspaper4k==0.9.3.1
nltk==3.9.1
numpy==2.2.4
openai==1.68.2
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
packaging==26.0
pandas==2.2.3
peewee==3.17.9
pgvector==0.4.0
pillow==11.1.0
platformdirs==4.3.7
primp==0.14.0
protobuf==6.33.5
psycopg==3.2.6
psycopg-binary==3.2.6
pycountry==24.6.1
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.8.1
pygments==2.19.1
pyjwt==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
referencing==0.37.0
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
rich==13.9.4
rich-toolkit==0.13.2
rpds-py==0.30.0
s3transfer==0.11.4
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
sqlalchemy==2.0.39
sse-starlette==3.0.3
starlette==0.46.1
tldextract==5.2.0
tomli==2.2.1
tqdm==4.67.1
typer==0.15.2
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.22.1
watchfiles==1.0.4
websockets==15.0.1
yfinance==0.2.55
zipp==3.23.0
